Look up Cantons from a separate API call
The remote API is currently returning an ID of zero for Canton in each
membership record.

We can work around this, and also get better names for the Cantons, by
looking them up at the Canton API endpoint.

In an ideal world, we'd rewrite this whole scraper using Scraped::JSON,
but let's start by just implementing the new functionality using this,
and evolve to it over time (I suspect we're going to want to remap the
Factions using the same approach Real Soon Now anyway)